refactor(fe): co-locate attribute key helpers with the channel handler
`extractAttributeName` and `extractScope` are attribute-key string parsers,
not store concerns. Move both into `channelHandlers/attributes.ts` — their
handler callsite for `extractAttributeName` and a shared util for the
views' `extractScope`. Nothing in `attributeConsent.store.ts` referenced
them any more.
