+++ src/frontend/src/lib/stores/attributeConsent.store.ts
@@ -56,17 +56,3 @@ export const attributeConsentResultStore: Readable<
 > = {
   subscribe: consentInternal.subscribe,
 };
-
-/** Extract the attribute name from a fully scoped key.
- *  e.g., "openid:https://accounts.google.com:email" → "email" */
-export const extractAttributeName = (key: string): string => {
-  const lastColon = key.lastIndexOf(":");
-  return lastColon >= 0 ? key.slice(lastColon + 1) : key;
-};
-
-/** Extract the scope from a fully scoped key.
- *  e.g., "openid:https://accounts.google.com:email" → "openid:https://accounts.google.com" */
-export const extractScope = (key: string): string | undefined => {
-  const lastColon = key.lastIndexOf(":");
-  return lastColon >= 0 ? key.slice(0, lastColon) : undefined;
-};
